+++ src/main/java/net/soulsweaponry/client/hud/FrostHudOverlay.java
@@ -1,0 +1,5 @@
+package net.soulsweaponry.client.hud;
+
+public class FrostHudOverlay {
+    //TODO implement and add textures (both before trigger/with and without frost cooldown active)
+}