+++ src/main/java/net/soulsweaponry/entity/effect/Freezing.java
@@ -5,6 +5,8 @@ import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.registry.tag.EntityTypeTags;
+import net.soulsweaponry.config.ConfigConstructor;
+import net.soulsweaponry.entitydata.FrostData;
 import net.soulsweaponry.items.axe.LeviathanAxe;
 import net.soulsweaponry.registry.EffectRegistry;
 import net.soulsweaponry.util.IAnimatedDeath;
@@ -27,6 +29,7 @@ public class Freezing extends StatusEffect {
         if (entity.getType().isIn(EntityTypeTags.FREEZE_IMMUNE_ENTITY_TYPES) || entity.hasStatusEffect(EffectRegistry.FROST_MOON)) {
             return false;
         }
+        FrostData.addFrost(entity, (int) (ConfigConstructor.permafrost_effect_base_frost_buildup + (ConfigConstructor.permafrost_effect_buildup_per_amp + 1) * amplifier));
         entity.setInPowderSnow(true);
         entity.setFrozenTicks(Math.min(entity.getMinFreezeDamageTicks(), ticks + amplifier));
         if (!entity.getWorld().isClient) {
@@ -35,11 +38,11 @@ public class Freezing extends StatusEffect {
         if (entity.isDead()) {
             if (entity instanceof IAnimatedDeath animated) {
                 if (animated.getDeathTicks() < 2) {
-                    LeviathanAxe.iceExplosion(entity.getWorld(), entity.getBlockPos(), entity.getAttacker(), amplifier);
+                    LeviathanAxe.iceExplosion(entity.getWorld(), entity.getBlockPos(), entity.getAttacker(), (amplifier + 1) * 1.5f, amplifier);
                 }
             }
             else if (entity.deathTime < 2) {
-                LeviathanAxe.iceExplosion(entity.getWorld(), entity.getBlockPos(), entity.getAttacker(), amplifier);
+                LeviathanAxe.iceExplosion(entity.getWorld(), entity.getBlockPos(), entity.getAttacker(), (amplifier + 1) * 1.5f, amplifier);
             }
         }
         return true;