+++ src/main/java/net/soulsweaponry/registry/AttributeRegistry.java
@@ -18,19 +18,25 @@ public class AttributeRegistry {
     public static RegistryEntry<EntityAttribute> BASE_POSTURE_INCREASE;
     public static RegistryEntry<EntityAttribute> BLEED_BUILDUP_RESISTANCE;
     public static RegistryEntry<EntityAttribute> BLEED_DAMAGE_RESISTANCE;
+    public static RegistryEntry<EntityAttribute> FROST_BUILDUP_RESISTANCE;
+    public static RegistryEntry<EntityAttribute> FROST_DAMAGE_RESISTANCE;
 
     public static void init() {
         POSTURE_BUILDUP_RESISTANCE = register("posture_buildup_resistance", 0.0, -10000.0, 10000.0);
         BASE_POSTURE_INCREASE = register("base_posture_increase", 0.0, -10000.0, 10000.0);
         BLEED_BUILDUP_RESISTANCE = register("bleed_buildup_resistance", 0.0, -10000.0, 10000.0);
         BLEED_DAMAGE_RESISTANCE = register("bleed_damage_resistance", 0.0, -10000.0, 10000.0);
+        FROST_BUILDUP_RESISTANCE = register("frost_buildup_resistance", 0.0, -10000.0, 10000.0);
+        FROST_DAMAGE_RESISTANCE = register("frost_damage_resistance", 0.0, -10000.0, 10000.0);
 
         DefaultAttributeContainer.Builder playerAttrs =
                 PlayerEntity.createPlayerAttributes()
                         .add(AttributeRegistry.POSTURE_BUILDUP_RESISTANCE)
                         .add(AttributeRegistry.BASE_POSTURE_INCREASE)
                         .add(AttributeRegistry.BLEED_BUILDUP_RESISTANCE)
-                        .add(AttributeRegistry.BLEED_DAMAGE_RESISTANCE);
+                        .add(AttributeRegistry.BLEED_DAMAGE_RESISTANCE)
+                        .add(AttributeRegistry.FROST_BUILDUP_RESISTANCE)
+                        .add(AttributeRegistry.FROST_DAMAGE_RESISTANCE);
         FabricDefaultAttributeRegistry.register(EntityType.PLAYER, playerAttrs);
     }